+++ src/client.rs
@@ -4,8 +4,12 @@ pub mod diagnostics;
 use std::io::Cursor;
 
 use bevy::{prelude::*, utils::EntityHashMap};
-use bevy_renet::{client_connected, renet::Bytes};
-use bevy_renet::{renet::RenetClient, transport::NetcodeClientPlugin, RenetClientPlugin};
+use bevy_renet::{
+    client_connected, client_just_disconnected,
+    renet::{Bytes, RenetClient},
+    transport::NetcodeClientPlugin,
+    RenetClientPlugin,
+};
 use bincode::{DefaultOptions, Options};
 use varint_rs::VarintReader;
 
@@ -33,13 +37,13 @@ impl Plugin for ClientPlugin {
                 PreUpdate,
                 ClientSet::Reset
                     .after(NetcodeClientPlugin::update_system)
-                    .run_if(bevy_renet::client_just_disconnected()),
+                    .run_if(client_just_disconnected()),
             )
             .configure_sets(
                 PreUpdate,
                 ClientSet::ResetEvents
                     .after(NetcodeClientPlugin::update_system)
-                    .run_if(bevy_renet::client_just_disconnected()),
+                    .run_if(client_just_disconnected()),
             )
             .configure_sets(
                 PostUpdate,
@@ -484,14 +488,14 @@ pub enum ClientSet {
     Send,
     /// Systems that reset the client.
     ///
-    /// Runs in `PreUpdate`.
+    /// Runs in `PreUpdate` when the client just disconnected.
     ///
     /// If this set is disabled, then you need to manually clean up the client after a disconnect or when
     /// reconnecting.
     Reset,
     /// Systems that reset queued server events.
     ///
-    /// Runs in `PreUpdate`.
+    /// Runs in `PreUpdate` when the client just disconnected.
     ///
     /// This is a separate set from `ClientSet::Reset` in case you want to enable/disable it separately.
     /// In general it is best practice to clear queued server events after a disconnect, whereas you may want to