+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor `ServerEventQueue` to have an explicit interface.
+- Cleanup all client events on disconnect.
 - Moved server event reset system to new set `ClientSet::ResetEvents` in `PreUpdate`.
 
 ## [0.19.0] - 2024-01-07